# docs/hosted-gateway-openapi.yaml
openapi: 3.1.0
info:
  title: ARC Chat Hosted Gateway
  version: 0.5.0
  description: Authenticated project and workspace API. Provider mutations require an approved server-side adapter.
servers:
  - url: https://gateway.example.invalid
    description: Replace with the operator-approved HTTPS origin.
security:
  - sessionCookie: []
paths:
...
  /api/v1/session:
    get:
      summary: Read authenticated session and CSRF state
      responses:
        "200":
          description: Session snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionSnapshot" }
        "401": { $ref: "#/components/responses/ApiError" }
...
  /api/v1/workspaces:
    get:
      summary: List authorized workspaces
      parameters:
        - in: query
          name: project_id
          schema: { $ref: "#/components/schemas/ProjectId" }
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Workspace" } }
    post:
      summary: Create a browser workspace record
      description: This records a link to static JupyterLite; it does not provision a remote kernel.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreateRequest" }
      responses:
        "201":
          description: Browser workspace record created
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace: { $ref: "#/components/schemas/Workspace" }
        "409": { $ref: "#/components/responses/ApiError" }
        "503": { $ref: "#/components/responses/ApiError" }
  /api/v1/workspaces/resolve:
    post:
      summary: Resolve resources persisted by a trusted provider adapter
      description: The client cannot submit provider candidates. Resolution needs the session CSRF header.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolutionRequest" }
      responses:
        "200":
          description: Provider-neutral resolution decision
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ResolutionDecision" }
        "400": { $ref: "#/components/responses/ApiError" }
...
  /api/v1/events:
    get:
      summary: Stream authorized project events
      parameters:
        - in: header
          name: Last-Event-ID
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Server-sent events; reconnect with the last event sequence.
          content:
            text/event-stream:
              schema: { type: string }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-arcchat-oauth
  parameters:
...
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, minLength: 16, maxLength: 128, pattern: "^[A-Za-z0-9_-]+$" }
  responses:
    ApiError:
      description: Structured API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
    ProjectId:
      type: string
      pattern: "^prj_[0-9a-f]{32}$"
    WorkspaceId:
      type: string
      pattern: "^ws_[0-9a-f]{32}$"
    SessionSnapshot:
      type: object
      required: [authenticated, user, projects, capabilities, csrf_token]
      properties:
        authenticated: { const: true }
        user:
          type: object
          properties:
            id: { type: string, pattern: "^usr_[0-9a-f]{32}$" }
            display_name: { type: string }
        projects: { type: array, items: { $ref: "#/components/schemas/ProjectId" } }
        capabilities:
          type: object
          additionalProperties: { type: boolean }
        csrf_token: { type: string }
...
    WorkspaceCreateRequest:
      type: object
      required: [project_id, provider_id]
      additionalProperties: false
      properties:
        project_id: { $ref: "#/components/schemas/ProjectId" }
        provider_id: { const: browser }
        application_id: { type: string, pattern: "^app_[0-9a-f]{32}$" }
        display_name: { type: string, minLength: 1, maxLength: 160 }
        reviewed_request_id: { type: string }
    Workspace:
      type: object
      required: [id, project_id, provider_id, state]
      properties:
        id: { $ref: "#/components/schemas/WorkspaceId" }
        project_id: { $ref: "#/components/schemas/ProjectId" }
        provider_id: { type: string }
        state: { type: string }
        display_name: { type: string }
        launch_url: { type: string, format: uri }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        metadata: { type: object, additionalProperties: { type: [string, integer, number, boolean, "null"] } }
    ResolutionRequest:
      type: object
      required: [project_id]
      additionalProperties: false
      properties:
        project_id: { $ref: "#/components/schemas/ProjectId" }
        workspace_id: { $ref: "#/components/schemas/WorkspaceId" }
        provider_id: { type: string }
        expectation:
          type: object
          additionalProperties: false
          properties:
            workspace_id: { $ref: "#/components/schemas/WorkspaceId" }
            allocation: { type: string }
            application_type: { type: string }
            job_name: { type: string }
            expected_state: { type: string }
            cluster: { type: string }
            provider: { type: string }
            created_after: { type: string, format: date-time }
    ResolutionDecision:
      type: object
      required: [action, resource_id, confidence, reason, candidates]
      properties:
        action: { type: string, enum: [reuse, wait, choose, start] }
        resource_id: { type: string }
        confidence: { type: string }
        reason: { type: string }
        evidence: { type: array, items: { type: string } }
        candidates: { type: array, items: { $ref: "#/components/schemas/ResourceCandidate" } }
    ResourceCandidate:
      type: object
      properties:
        resource_id: { type: string, pattern: "^prsrc_[0-9a-f]{32}$" }
        provider_resource_id: { type: string, pattern: "^prsrc_[0-9a-f]{32}$" }
        state: { type: string }
        name: { type: string }
        provider: { type: string }
        kind: { type: string }
        project_id: { $ref: "#/components/schemas/ProjectId" }
        workspace_id: { type: string }
        created_at: { type: string }
        cluster: { type: string }
...
    ApiError:
      type: object
      required: [error_version, code, message, request_id, recovery_action]
      properties:
        error_version: { const: 1 }
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        recovery_action: { type: string }